requirements.txt
langchain-core
langchainhub
langchain-community
langchain-google-genai
langchain-pinecone
python-dotenv